Some warning fixes

+++ Editor/main_window/plugins/pge_editorpluginmanager.cpp
@@ -11,12 +11,12 @@ bool PGE_EditorPluginManager::hasPlugins() const
 
 unsigned int PGE_EditorPluginManager::getPluginCount() const
 {
-    return m_loadedPlugins.size();
+    return uint(m_loadedPlugins.size());
 }
 
-const PGE_EditorPluginItem &PGE_EditorPluginManager::getPluginInfo(int index) const
+const PGE_EditorPluginItem &PGE_EditorPluginManager::getPluginInfo(unsigned int index) const
 {
-    return m_loadedPlugins[index];
+    return m_loadedPlugins[int(index)];
 }
 
 void PGE_EditorPluginManager::loadPluginsInDir(const QDir &dir)
@@ -32,7 +32,7 @@ void PGE_EditorPluginManager::loadPluginsInDir(const QDir &dir)
         QDir nextDir = dir;
         nextDir.cd(subFolder);
 
-        if(nextDir.exists("main.js")){
+        if(nextDir.exists("main.js")) {
             LogDebug(QString("Found ") + subFolder + " and attempt to load main.js!");
             bool ok = false;
             const QVariantMap result = m_engine.loadFileByExpcetedResult<QVariantMap>(nextDir.canonicalPath() + "/main.js", &ok);
@@ -54,7 +54,7 @@ void PGE_EditorPluginManager::loadPluginsInDir(const QDir &dir)
             m_loadedPlugins.push_back(PGE_EditorPluginItem(std::move(pluginName), std::move(authorName), std::move(description), version));
 
             WriteToLog(PGE_LogLevel::Debug, "Successfully loaded plugin\"" + pluginName + "\"!");
-        }else{
+        } else {
             LogDebug("Found package, but not main.js!")
         }
     }
